ai-service/requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
spacy
pytesseract
pymupdf
requests
python-multipart
pillow
python-dotenv
